refactor expenses_adapter_spec
Extract:-
* test setup into fixtures
* shared examples into support
